refactor: clean up and rename actionCreators.js to actions.js
